requirements.txt
Django
djangorestframework
psycopg2
Pillow
requests
pyzbar

